+++ src/components/table-ajax/definition.js
@@ -6,8 +6,8 @@ import tableCellDefinition from './../table/table-cell/definition';
 import tableHeaderDefinition from './../table/table-header/definition';
 import tableSubheaderDefinition from './../table/table-subheader/definition';
 
-let definition = new Definition('table-ajax', TableAjax, {
-  description: `Ajax control: A table of relational data to view or interact with.`,
+const definition = new Definition('table-ajax', TableAjax, {
+  description: 'Ajax control: A table of relational data to view or interact with.',
   designerNotes: `
 * This control is the same as [Table](/components/table), but uses Ajax.
 * Ajax loads data from a specified source as needed, rather than data in the page markup.
@@ -20,13 +20,18 @@ let definition = new Definition('table-ajax', TableAjax, {
     onChange: ComponentActions.updateTableAjax
   },
   propTypes: {
-    path: "String"
+    formatData: 'Function',
+    path: 'String'
   },
   propDescriptions: {
-    path: "The path to make XHR requests to."
+    formatData: 'Callback function for formatting the data received via Ajax requests into the format required by the table',
+    path: 'The path to make XHR requests to.'
   },
   requiredProps: ['path'],
-  hiddenProps: ['path']
+  hiddenProps: [
+    'formatData',
+    'path'
+  ]
 });
 
 definition.isATable();